update token screen structure
